requirements.txt
pandas==2.1.0
numpy==1.24.0
requests==2.31.0
python-dotenv==1.0.0
supabase==1.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
xgboost==2.0.0
plotly==5.17.0 